feat: combine prefab and scene parse
